+++ GenericModConfigMenu/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed mouse-wheel-scroll sound playing even after the menu is closed.
+* Fixed API `TryGetCurrentMenu` method not working when on the title screen.
 * Improved translations. Thanks to BuslaevLegat (added Russian)!
 
 ## 1.7.0